src/csoapi/requirements.txt
django
djangorestframework
psycopg2
geopandas
requests
pandas
lxml
django-redis
hiredis
django-webpack-loader